Add files via upload
This is an allocation model problem, each column represents a television show in which cereal commercials are aired and each row is a gender/age demographic group liked, or like to eat! This case, helped us to understand the business of producing cereal and the role of advertising in this business.